rename: hook 및 type service 등 모든 파일이 디렉토리 위치 수정해야함에 따라 원복
