+++ src/pages/SignUp/hooks/useSignUpQuery.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import { getSingUp } from '../../api';
+import { getSingUp } from '../../../api';
 import { convertCheckboxQuery, convertSkillQuery } from '../service/convertSignUpQuery';
 
 const SKILL_DEPTH_THREE_LIST = [
